refactor: relocate Section component
