+++ src/components/section/Section.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
-import Section from "./Section";
+import Section from "./Section.tsx";
 
 describe("Section component", () => {
   it("renders title", () => {
